Module anymap::any

The different types of Any for use in a map.

This stuff is all based on std::any, but goes a little further, with CloneAny being a cloneable Any and with the Send and Sync bounds possible on both Any and CloneAny.

Traits

Any
See also CloneAny for a cloneable version of this trait.

CloneAny
See also Any for a version without the Clone requirement.

UncheckedAnyExt