Module anymap::raw

The raw form of a Map, allowing untyped access.

All relevant details are in the RawMap struct.

Structs

Drain
RawMap drain iterator.

IntoIter
RawMap move iterator.

Iter
RawMap iterator.

IterMut
RawMap mutable iterator.

OccupiedEntry
A view into a single occupied location in a RawMap.

RawMap
The raw, underlying form of a Map.

VacantEntry
A view into a single empty location in a RawMap.

Enums

Entry
A view into a single location in a RawMap, which may be vacant or occupied.