Enum anymap::raw::Entry

pub enum Entry<'a, A: ?Sized + UncheckedAnyExt> {
    Occupied(OccupiedEntry<'a, A>),
    Vacant(VacantEntry<'a, A>),
}

A view into a single location in a RawMap, which may be vacant or occupied.

Variants

Occupied

An occupied Entry

Vacant

A vacant Entry

Methods

impl<'a, A: ?Sized + UncheckedAnyExt> Entry<'a, A>

unsafe fn or_insert(self, default: Box<A>) -> &'a mut A

Ensures a value is in the entry by inserting the default if empty, and returns a mutable reference to the value in the entry.

It is the caller’s responsibility to ensure that the key of the entry corresponds with the type ID of value. If they do not, memory safety may be violated.

unsafe fn or_insert_with<F: FnOnce() -> Box<A>>(self, default: F) -> &'a mut A

Ensures a value is in the entry by inserting the result of the default function if empty, and returns a mutable reference to the value in the entry.

It is the caller’s responsibility to ensure that the key of the entry corresponds with the type ID of value. If they do not, memory safety may be violated.