Struct anymap::raw::OccupiedEntry

pub struct OccupiedEntry<'a, A: ?Sized + UncheckedAnyExt> {
    // some fields omitted
}

A view into a single occupied location in a RawMap.

Methods

impl<'a, A: ?Sized + UncheckedAnyExt> OccupiedEntry<'a, A>

fn get(&self) -> &A

Gets a reference to the value in the entry.

fn get_mut(&mut self) -> &mut A

Gets a mutable reference to the value in the entry.

fn into_mut(self) -> &'a mut A

Converts the OccupiedEntry into a mutable reference to the value in the entry with a lifetime bound to the collection itself.

unsafe fn insert(&mut self, value: Box<A>) -> Box<A>

Sets the value of the entry, and returns the entry's old value.

It is the caller’s responsibility to ensure that the key of the entry corresponds with the type ID of value. If they do not, memory safety may be violated.

fn remove(self) -> Box<A>

Takes the value out of the entry, and returns it.