Struct anymap::AnyMap

pub struct AnyMap {
    // some fields omitted
}

A map containing zero or one values for any given type and allowing convenient, type-safe access to those values.

let mut data = AnyMap::new();
assert_eq!(data.find(), None::<&int>);
data.insert(42i);
assert_eq!(data.find(), Some(&42i));
data.remove::<int>();
assert_eq!(data.find::<int>(), None);

#[deriving(PartialEq, Show)]
struct Foo {
    str: String,
}

assert_eq!(data.find::<Foo>(), None);
data.insert(Foo { str: "foo".to_string() });
assert_eq!(data.find(), Some(&Foo { str: "foo".to_string() }));
data.find_mut::<Foo>().map(|foo| foo.str.push('t'));
assert_eq!(data.find::<Foo>().unwrap().str.as_slice(), "foot");

Values containing non-static references are not permitted.

Methods

impl AnyMap

fn new() -> AnyMap

Construct a new AnyMap.

impl AnyMap

fn find<'a, T: 'static>(&'a self) -> Option<&'a T>

Retrieve the value stored in the map for the type T, if it exists.

fn find_mut<'a, T: 'static>(&'a mut self) -> Option<&'a mut T>

Retrieve a mutable reference to the value stored in the map for the type T, if it exists.

fn insert<T: 'static>(&mut self, value: T)

Set the value contained in the map for the type T. This will override any previous value stored.

fn remove<T: 'static>(&mut self)

Remove the value for the type T if it existed.

fn contains<T: 'static>(&self) -> bool

Does a value of type T exist?